parser: fixing tests to work with the new parser
